requirements.txt
picamera2
pillow
escpos